// src/app/models/IControlLoteMateriales.ts
export interface IControlLoteMateriales {
  idControlLoteMateriales?: number | null;
  codigoModelo?: string | null;
  codigoPautas?: string | null;
  numeroOp?: string | null;
  cantidad?: number | null;
  createdDate?: string | null;
  lastModifiedDate?: string | null;
  nombreSupervisor?: string | null;
  deleted?: boolean;
  idControlLote?: number | null;
  descripcion?: string | null;
}